Add files via upload
Realized I have been uploading things incorrectly/not uploading to GitHub (thought the repositories synced?). 

Attached are three files for the PyBank Assignment, I believe main.py and 'main.py'ipynb are the same files (actual code to execute print outs) and the New_PyBank is the text file we were asked to product to show the printed output. May have confused these (if so. apologies)!